git-svn-id: https://svn.apache.org/repos/asf/poi/branches/hemf@1843032 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/ss/formula/atp/RandBetween.java
@@ -77,7 +77,7 @@ final class RandBetween implements FreeRefFunction{
 			top = bottom;
 		}
 		
-		return new NumberEval((bottom + (int)(Math.random() * ((top - bottom) + 1))));
+		return new NumberEval((bottom + (long)(Math.random() * ((top - bottom) + 1))));
 		
 	}
